separate data sources for dialog (single item) and table (all items of a dictionary)
