+++ ui.apps/src/main/content/jcr_root/apps/aem-dictionary-translator/content/granite/dialog/dictionary/key/update/.content.xml
@@ -33,7 +33,7 @@
                                             sling:resourceType="granite/ui/components/coral/foundation/form/responses/reload"/>
                                     <datasource
                                             jcr:primaryType="nt:unstructured"
-                                            sling:resourceType="aem-dictionary-translator/datasource/combining-message-entry"/>
+                                            sling:resourceType="aem-dictionary-translator/datasource/combining-message-entry-for-dialog"/>
                                 </form>
                             </items>
                         </column>
